Fix scan_runs counting
Addition of the background scanning mechanism in commit
60b94c9819199c0427e623080ebae27fe44be6b5 moved the scan trigger
into a new function that was also incrementing the scan_runs
counter, but the removal of the previous scan_runs incrementation
was forgotten from that patch. This counter should only be updated
into a single location, so remove the old one. This improves AP
selection for WPS provisioning by not skipping some of the initial
scans.